fix: Create documentation for both approaches

+++ tools/start-gitlab-runner-podman.sh
@@ -1,9 +1,66 @@
 #!/usr/bin/env bash
 # shellcheck disable=SC1090,SC1091,SC2015
 #
-# Create a gitlab runner by first visiting:
-# `CI/CD Settings page and creating a `linux` runner.
-# The token has to be given to this function.
+# Create a gitlab runner (docker executor)
+# by first visiting `CI/CD Settings` page and
+# creating a `linux` runner which gives you a `<token>` needed
+# for this script.
+#
+# This creates a container which runs the Gitlab runner
+# which will execute jobs over the `podman` executor.
+#
+# This container is based on [`pipglr`](https://gitlab.com/qontainers/pipglr)
+# which uses two container volumes (`pipglr-storage` and `pipglr-cache`)
+# which are attached and contains
+# `podman` (user `podman`) and `gitlab-runner` (user `runner`)
+# to have a rootless container experience which provides much more security.
+# The two volumes contain the images created/built by CI. The volumes can safely
+# be wiped if space is needed.
+#
+# The podman socket created inside this container
+# will be mounted to each job container the `gitlab-runner` creates.
+# This makes also use of caching directly possible which is cool.
+#
+# Usage:
+# ```shell
+# start-gitlab-runner-docker.sh [--force] [<token>]
+# ```
+# Read token from stdin.
+# ```shell
+# start-gitlab-runner-docker.sh [--force] -
+#
+# Usage in Pipeline:
+#
+# A job which uses `podman` (linked to podman) to run/build images.
+# The gitlab-runner cannot serve `services` statements as in the
+# `start-gitlab-runner-docker.sh` (uses `docker`
+# `--link` which is anyway deprecated)
+#
+# ```yaml
+# podman-remote-run-build:
+#   image: quay.io/podman/stable:latest
+#   variables:
+#     CONTAINER_HOST: unix://var/run/docker.sock
+#   script:
+#     - podman info
+#     - podman run alpine:latest cat /etc/os-release
+#     - podman build -f Dockerfile .
+# ```
+#
+# The following (custom build image) also works:
+#
+# ```yaml
+# podman-remote-alpine-run-build:
+#   image: alpine:latest
+#   variables:
+#     CONTAINER_HOST: unix://var/run/docker.sock
+#   script:
+#     - apk add podman
+#     - podman info
+#     - podman run alpine:latest cat /etc/os-release
+#     - podman build -f Dockerfile .
+# ```
+
 set -e
 set -u
 
@@ -19,8 +76,7 @@ image="registry.gitlab.com/qontainers/pipglr:latest@sha256:ca3093e580bd617d5f92b
 
 function clean_up() {
     if [ -f "$config_dir/config.toml" ]; then
-        true
-        # rm -rf "$config_dir/config.toml"
+        rm -rf "$config_dir/config.toml"
     fi
 }
 
@@ -30,7 +86,7 @@ function modify_config() {
     local value="$2"
     local type="${3:-json}"
 
-    podman run -v "$config_dir/config.toml:/config.toml" \
+    podman run --rm -v "$config_dir/config.toml:/config.toml" \
         "ghcr.io/tomwright/dasel" put -f /config.toml \
         -t "$type" \
         -s "$key" \
@@ -40,20 +96,29 @@ function modify_config() {
 
 function register_runner() {
     print_info "Registering gitlab-runner ..."
-    local token="$1"
+    local token="${1:-}"
+
+    if [ "$token" = "-" ] || [ -z "$token" ]; then
+        read -rs -p "Enter Gitlab Runner Token: " token ||
+            die "Could not read token from stdin."
+    fi
 
     podman secret rm REGISTRATION_TOKEN &>/dev/null || true
     echo "$token" | podman secret create REGISTRATION_TOKEN - ||
         die "Could not set registration token secret."
 
+    # Register Gitlab runner.
     (cd "$config_dir" &&
         touch config.toml &&
         podman container runlabel register "$image") ||
         die "Could not register gitlab-runner."
 
+    # Modify Gitlab runner config.
     modify_config ".concurrent" "$max_jobs"
-    modify_config ".runners.first().docker.pull_policy" '["always", "if-not-present"]'
-    modify_config ".runners.first().docker.volumes.append()" "/home/runner/podman.sock:/var/run/docker.sock:rw" string
+    modify_config ".runners.first().docker.pull_policy" \
+        '["always", "if-not-present"]'
+    modify_config ".runners.first().docker.volumes.append()" \
+        "/home/runner/podman.sock:/var/run/docker.sock:rw" string
 
     podman secret rm config.toml &>/dev/null || true
     podman secret create config.toml "$config_dir/config.toml" ||
@@ -100,12 +165,10 @@ function start_runner() {
 }
 
 function create() {
-    local token="${1:?First argument must be the runner token.}"
-
     rm -rf "$config_dir" >/dev/null || true
     mkdir -p "$config_dir"
 
-    register_runner "$token"
+    register_runner "$@"
     assert_volumes
 
     start_runner
@@ -116,13 +179,21 @@ function stop() {
         print_info "Stop runner '$runner_name' ..."
         podman stop "$runner_name"
 
+    fi
+
+    if is_exited; then
         # shellcheck disable=SC2046
         podman rm $(podman ps -a -q)
     fi
 }
 
 function is_running() {
-    [ "$(podman inspect -f '{{.State.Running}}' "$runner_name" 2>/dev/null || true)" = 'true' ] || return 1
+    [ "$(podman inspect -f '{{.State.Status}}' "$runner_name" 2>/dev/null || true)" = 'running' ] || return 1
+    return 0
+}
+
+function is_exited() {
+    [ "$(podman inspect -f '{{.State.Status}}' "$runner_name" 2>/dev/null || true)" = 'exited' ] || return 1
     return 0
 }
